dashboard/backend/requirements.txt
fastapi
uvicorn
plaid-python
requests
pandas
